+++ packages/amplify-prompts/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-prompts@2.6.6-rc.e9e0d157d.0...amplify-prompts@2.6.6) (2023-03-23)
+
+**Note:** Version bump only for package amplify-prompts
+
+
+
+
+
+## [2.6.6-rc.e9e0d157d.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-prompts@2.6.5...amplify-prompts@2.6.6-rc.e9e0d157d.0) (2023-03-17)
+
+**Note:** Version bump only for package amplify-prompts
+
+
+
+
+
 ## [2.6.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-prompts@2.6.4...amplify-prompts@2.6.5) (2023-03-13)